Don't insert padding between lists in requests
Lists of types that have the same padding as each other should not have
any padding automatically inserted by libxcb.

Fixes xcb_randr_set_crtc_gamma when "size" is an odd number.

Requires us to loop over all the fields, not just the param fields, so
that we can generate code for <pad align> (which was previously missing,
but went unnoticed because libxcb was always behaving as if a <pad
align="4" /> was present between each list).

Incidentally reduces the number of unnecessary pad entries in the
xcb_parts lists (eg. between lists when the first is a list of CARD32).

Signed-off-by: Peter Harris <pharris2@rocketsoftware.com>